requirements.txt
python-dotenv
pyyaml
supabase
